#101 - 타이머 셋하면 알람 자동 설정 구현
