+++ app/src/main/java/com/nexters/ticktock/alarmsetting/AlarmSettingThirdActivity.kt
@@ -1,8 +1,12 @@
 package com.nexters.ticktock.alarmsetting
 
+import android.app.AlarmManager
+import android.app.PendingIntent
+import android.content.Context
 import android.content.Intent
 import android.databinding.DataBindingUtil
 import android.graphics.Typeface
+import android.os.Build
 import android.os.Bundle
 import android.support.v4.content.ContextCompat
 import android.support.v7.app.AppCompatActivity
@@ -25,6 +29,7 @@ import com.nexters.ticktock.model.Alarm
 import com.nexters.ticktock.model.Step
 import com.nexters.ticktock.model.enums.Day
 import com.nexters.ticktock.model.enums.TickTockColor
+import com.nexters.ticktock.timer.AlarmReceiver
 import com.nexters.ticktock.timer.TimerActivity
 import com.nexters.ticktock.utils.Time
 import com.nexters.ticktock.utils.getHighlightedString
@@ -139,6 +144,38 @@ class AlarmSettingThirdActivity : AppCompatActivity(), View.OnClickListener, Rad
             binding.layoutSave.id -> {
                 // TODO 저장
                 saveData()
+
+                val alarmDao = TickTockDBHelper.getInstance(this).alarmDao
+
+                val calendarSet = Calendar.getInstance()
+                val startHour = startTime.time / 60
+                val startMinute = startTime.minute
+
+                calendarSet.set(Calendar.HOUR_OF_DAY, startHour)
+                calendarSet.set(Calendar.MINUTE, startMinute)
+
+                Log.d("StartAlarm", "$startHour:$startMinute")
+
+                val mAlarmIntent:Intent = Intent(this, AlarmReceiver::class.java)
+                val pIntent : PendingIntent = PendingIntent.getBroadcast(this, alarmDao.findAll().size, mAlarmIntent, PendingIntent.FLAG_UPDATE_CURRENT)
+
+                val alarmManager : AlarmManager = getSystemService(Context.ALARM_SERVICE) as AlarmManager
+
+                val rightNow : Calendar = Calendar.getInstance()
+
+                if(rightNow.timeInMillis < calendarSet.timeInMillis) {
+
+                    alarmManager.set(AlarmManager.RTC_WAKEUP, calendarSet.timeInMillis, pIntent)
+                    if(Build.VERSION.SDK_INT >= 23)
+                        alarmManager.setExactAndAllowWhileIdle(AlarmManager.RTC_WAKEUP, calendarSet.timeInMillis, pIntent)
+
+                    else if(Build.VERSION.SDK_INT >= 21)
+                        alarmManager.setExact(AlarmManager.RTC_WAKEUP, calendarSet.timeInMillis, pIntent)
+
+                    else
+                        alarmManager.set(AlarmManager.RTC_WAKEUP, calendarSet.timeInMillis, pIntent)
+                }
+
                 val intent = Intent(this, CardActivity::class.java)
                 startActivity(intent)
             }
